[Bronze III] Title: 악마의 제안, Time: 176 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/23972. 악마의 제안/악마의 제안.cc
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    long long int k, n;
+    int k, n;
     long long int x;
     cin >> k >> n;
    long long int i = 1;
